backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlmodel==0.0.14
psycopg2-binary==2.9.9
langchain==0.1.0
langchain-google-genai>=1.0.0
google-generativeai>=0.3.0
langchain-core==0.1.10
celery==5.3.4
redis==5.0.1
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

